refactor: decompose permissions settings into focused modules
Extract reusable components and MCP tab logic from the 559-line
permissions.tsx into dedicated files:

- permissions/components.tsx: ToolRow, ToolsetRow, SectionCard, EmptyState
- permissions/mcp-tools-tab.tsx: McpToolsTab, useMcpToolsetGroups, filterMcpGroups
- permissions/types.ts: shared EditingTarget type

Consolidates the duplicated EditingTarget/PermissionTarget type into
a single source of truth used by both permissions.tsx and the
permission-policy-editor.
